refactor(tests): Organize tests in folders and files per contract
