+++ test/erc20guild/ERC20Guild.js
@@ -1,9 +1,12 @@
-import * as helpers from "./helpers";
-const constants = require("./helpers/constants");
-const ERC20Guild = artifacts.require("ERC20Guild.sol");
-const IERC20Guild = artifacts.require("IERC20Guild.sol");
-const ActionMock = artifacts.require("ActionMock.sol");
-const { fixSignature } = require("./helpers/sign");
+import * as helpers from "../helpers";
+const { fixSignature } = require("../helpers/sign");
+const {
+  createDAO,
+  createAndSetupGuildToken,
+  createProposal,
+  setAllVotesOnProposal
+} = require("../helpers/guild");
+
 const {
   BN,
   expectEvent,
@@ -13,17 +16,16 @@ const {
   ether,
   time
 } = require("@openzeppelin/test-helpers");
-const {
-  createDAO,
-  createAndSetupGuildToken,
-  createProposal,
-  setAllVotesOnProposal
-} = require("./helpers/guild");
+
+const ERC20Guild = artifacts.require("ERC20Guild.sol");
+const IERC20Guild = artifacts.require("IERC20Guild.sol");
+const ActionMock = artifacts.require("ActionMock.sol");
 
 require("chai").should();
 
 contract("ERC20Guild", function (accounts) {
   
+  const constants = helpers.constants;
   const ZERO = new BN("0");
   const TIMELOCK = new BN("60");
   const VOTE_GAS = new BN("50000"); // 50k
